+++ src/promptflow/promptflow/_cli/_pf_azure/_flow.py
@@ -34,6 +34,7 @@ def add_parser_flow(subparsers):
     )
     flow_subparsers = flow_parser.add_subparsers()
     add_parser_flow_create(flow_subparsers)
+    add_parser_flow_update(flow_subparsers)
     add_parser_flow_show(flow_subparsers)
     add_parser_flow_list(flow_subparsers)
     flow_parser.set_defaults(action="flow")
@@ -63,9 +64,9 @@ pfazure flow create --flow <flow-folder-path> --set display_name=<flow-display-n
         "--flow", type=str, help="Source folder of the flow."
     )
     add_params = [
-        _set_workspace_argument_for_subparsers,
         add_param_source,
         add_param_set,
+        _set_workspace_argument_for_subparsers,
     ] + base_params
 
     activate_action(
@@ -75,6 +76,43 @@ pfazure flow create --flow <flow-folder-path> --set display_name=<flow-display-n
         add_params=add_params,
         subparsers=subparsers,
         help_message="Create a flow to Azure with local flow folder.",
+        action_param_name="sub_action",
+    )
+
+
+def add_parser_flow_update(subparsers):
+    """Add flow update parser to the pf flow subparsers."""
+    epilog = """
+Use "--set" to set flow properties that you want to update. Supported properties are: [display_name, description, tags].
+
+Note:
+    1. In "--set" parameter, if the key name consists of multiple words, use snake-case instead of kebab-case. e.g. "--set display_name=<flow-display-name>"
+    2. Parameter flow is required to update a flow. It's a guid that can be found from 2 ways:
+        a. After creating a flow to azure, it can be found in the printed message in "name" attribute.
+        b. Open a flow in azure portal, the guid is in the url. e.g. https://ml.azure.com/prompts/flow/<workspace-id>/<flow-name>/xxx
+
+Examples:
+
+# Update a flow display name
+pfazure flow update --flow <flow-name> --set display_name=<flow-display-name>
+"""  # noqa: E501
+
+    add_param_source = lambda parser: parser.add_argument(  # noqa: E731
+        "--flow", type=str, help="Flow name to be updated which is a guid."
+    )
+    add_params = [
+        add_param_source,
+        add_param_set,
+        _set_workspace_argument_for_subparsers,
+    ] + base_params
+
+    activate_action(
+        name="update",
+        description="A CLI tool to update a flow's metadata on Azure.",
+        epilog=epilog,
+        add_params=add_params,
+        subparsers=subparsers,
+        help_message="Update a flow's metadata on azure.",
         action_param_name="sub_action",
     )
 
@@ -168,11 +206,13 @@ def add_parser_flow_download(subparsers):
 
 def dispatch_flow_commands(args: argparse.Namespace):
     if args.sub_action == "create":
-        create_flow(args)
+        create_or_update_flow(args)
     elif args.sub_action == "show":
         show_flow(args)
     elif args.sub_action == "list":
         list_flows(args)
+    elif args.sub_action == "update":
+        create_or_update_flow(args)
 
 
 def _get_flow_operation(subscription_id, resource_group, workspace_name):
@@ -180,7 +220,7 @@ def _get_flow_operation(subscription_id, resource_group, workspace_name):
     return pf_client._flows
 
 
-def create_flow(args: argparse.Namespace):
+def create_or_update_flow(args: argparse.Namespace):
     """Create a flow for promptflow."""
     pf = _get_azure_pf_client(args.subscription, args.resource_group, args.workspace_name, debug=args.debug)
     params = _parse_flow_metadata_args(args.params_override)